# docs/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spark Security Agent Backend",
    "description": "安全事件智能处置后端 MVP 接口文档。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/runs": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "启动安全事件处理主流程",
        "operationId": "start_event_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "查询安全事件详情",
        "operationId": "get_event",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/timeline": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "查询安全事件状态时间线",
        "operationId": "get_event_timeline",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimelineEntry"
                  },
                  "title": "Response Get Event Timeline"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/approval": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "提交安全事件处置审批结果",
        "operationId": "submit_event_approval",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalDecision": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "approver": {
            "type": "string",
            "title": "Approver"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "approved",
          "approver",
          "reason",
          "idempotency_key"
        ],
        "title": "ApprovalDecision"
      },
      "ApprovalStatus": {
        "type": "string",
        "enum": [
          "not_required",
          "pending",
          "approved",
          "rejected"
        ],
        "title": "ApprovalStatus"
      },
      "BusinessStatus": {
        "type": "string",
        "enum": [
          "RECEIVED",
          "CORRELATING",
          "TRIAGED",
          "INVESTIGATING",
          "DECISION_READY",
          "APPROVAL_REQUIRED",
          "EXECUTING",
          "VERIFYING",
          "COMPLETED",
          "HUMAN_REQUIRED",
          "FAILED"
        ],
        "title": "BusinessStatus"
      },
      "ErrorRecord": {
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "recoverable": {
            "type": "boolean",
            "title": "Recoverable",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "stage",
          "message"
        ],
        "title": "ErrorRecord"
      },
      "EventContext": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "2026-08-21.mvp.v1"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "status": {
            "$ref": "#/components/schemas/BusinessStatus"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "alert_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alert Refs"
          },
          "event_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SecurityEvent"
              },
              {
                "type": "null"
              }
            ]
          },
          "triage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriageResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "investigation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "response": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntry"
            },
            "type": "array",
            "title": "Timeline"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorRecord"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "run_id",
          "event_id",
          "status",
          "source"
        ],
        "title": "EventContext"
      },
...
      "ExecutionMode": {
        "type": "string",
        "enum": [
          "mock",
          "real"
        ],
        "title": "ExecutionMode"
      },
      "ExecutionResult": {
        "properties": {
          "executed": {
            "type": "boolean",
            "title": "Executed"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "platform_status": {
            "type": "string",
            "title": "Platform Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "executed",
          "mode",
          "platform_status",
          "idempotency_key"
        ],
        "title": "ExecutionResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationReport": {
        "properties": {
          "conclusion": {
            "$ref": "#/components/schemas/TruthVerdict"
          },
          "final_confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Final Confidence"
          },
          "timeline": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Timeline"
          },
          "tool_results": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tool Results"
          },
          "key_evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Evidence Refs"
          },
          "evidence_relations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Relations"
          },
          "affected_objects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Objects"
          },
          "unresolved_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unresolved Questions"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "needs_human": {
            "type": "boolean",
            "title": "Needs Human",
            "default": false
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/InvestigationStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "conclusion",
          "final_confidence",
          "summary"
        ],
        "title": "InvestigationReport"
      },
      "InvestigationStep": {
        "properties": {
          "step_no": {
            "type": "integer",
            "title": "Step No"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "tool_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "tool_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "observation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation"
          }
        },
        "type": "object",
        "required": [
          "step_no",
          "goal"
        ],
        "title": "InvestigationStep"
      },
...
      "Priority": {
        "type": "string",
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "Priority"
      },
      "ResponsePlan": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "risk_level": {
            "$ref": "#/components/schemas/ToolRiskLevel"
          },
          "approval_required": {
            "type": "boolean",
            "title": "Approval Required"
          },
          "rollback_available": {
            "type": "boolean",
            "title": "Rollback Available"
          }
        },
        "type": "object",
        "required": [
          "action",
          "target",
          "reason",
          "risk_level",
          "approval_required",
          "rollback_available"
        ],
        "title": "ResponsePlan"
      },
      "ResponseResult": {
        "properties": {
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponsePlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "execution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ResponseResult"
      },
      "SecurityEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "alert_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alert Refs"
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          },
          "entities": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Entities"
          },
          "correlation_reason": {
            "type": "string",
            "title": "Correlation Reason"
          },
          "alert_count_before": {
            "type": "integer",
            "title": "Alert Count Before"
          },
          "event_count_after": {
            "type": "integer",
            "title": "Event Count After"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "alert_refs",
          "first_seen_at",
          "last_seen_at",
          "correlation_reason",
          "alert_count_before",
          "event_count_after",
          "summary"
        ],
        "title": "SecurityEvent"
      },
      "StartRunRequest": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "fixed_sample",
              "jsonl_sample",
              "xdr"
            ],
            "title": "Source",
            "default": "fixed_sample"
          },
          "sample_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Id"
          },
          "xdr_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xdr Event Id"
          }
        },
        "type": "object",
        "title": "StartRunRequest"
      },
      "TimelineEntry": {
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "status": {
            "$ref": "#/components/schemas/BusinessStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "TimelineEntry"
      },
      "ToolCallStatus": {
        "type": "string",
        "enum": [
          "success",
          "failed",
          "partial_success"
        ],
        "title": "ToolCallStatus"
      },
      "ToolErrorType": {
        "type": "string",
        "enum": [
          "auth",
          "timeout",
          "validation",
          "unsupported_tool",
          "platform_error",
          "unknown"
        ],
        "title": "ToolErrorType"
      },
      "ToolRequest": {
        "properties": {
          "call_id": {
            "type": "string",
            "title": "Call Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "stage": {
            "$ref": "#/components/schemas/BusinessStatus"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "action_name": {
            "type": "string",
            "title": "Action Name"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "param_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Param Refs"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "risk_level": {
            "$ref": "#/components/schemas/ToolRiskLevel"
          },
          "approval_status": {
            "$ref": "#/components/schemas/ApprovalStatus",
            "default": "not_required"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "timeout_seconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 30
          },
          "attempt": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Attempt",
            "default": 1
          },
          "max_attempts": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Attempts",
            "default": 1
          },
          "sensitive_param_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sensitive Param Keys"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "event_id",
          "stage",
          "tool_name",
          "action_name",
          "params",
          "reason",
          "idempotency_key",
          "risk_level"
        ],
        "title": "ToolRequest"
      },
      "ToolResult": {
        "properties": {
          "call_id": {
            "type": "string",
            "title": "Call Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "action_name": {
            "type": "string",
            "title": "Action Name"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "status": {
            "$ref": "#/components/schemas/ToolCallStatus"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "raw_result_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Result Ref"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "output_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Output Refs"
          },
          "output_preview": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Preview"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "error_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolErrorType"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "platform_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Status"
          },
          "external_side_effect": {
            "type": "boolean",
            "title": "External Side Effect",
            "default": false
          },
          "side_effect_type": {
            "$ref": "#/components/schemas/ToolSideEffectType",
            "default": "none"
          },
          "attempt": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Attempt",
            "default": 1
          },
          "max_attempts": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Attempts",
            "default": 1
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ended At"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "call_id",
          "trace_id",
          "event_id",
          "tool_name",
          "action_name",
          "idempotency_key",
          "status",
          "summary",
          "started_at",
          "ended_at",
          "duration_ms"
        ],
        "title": "ToolResult"
      },
      "ToolRiskLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "ToolRiskLevel"
      },
      "ToolSideEffectType": {
        "type": "string",
        "enum": [
          "none",
          "read_only",
          "state_change"
        ],
        "title": "ToolSideEffectType"
      },
      "TriageResult": {
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/TruthVerdict"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "supporting_evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supporting Evidence Refs"
          },
          "opposing_evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Opposing Evidence Refs"
          },
          "evidence_gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Gaps"
          },
          "should_investigate": {
            "type": "boolean",
            "title": "Should Investigate"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "verdict",
          "confidence",
          "risk_score",
          "priority",
          "should_investigate",
          "summary"
        ],
        "title": "TriageResult"
      },
      "TruthVerdict": {
        "type": "string",
        "enum": [
          "malicious",
          "benign",
          "uncertain"
        ],
        "title": "TruthVerdict"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationResult": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "adjustment_suggestion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjustment Suggestion"
          },
          "final_status": {
            "$ref": "#/components/schemas/BusinessStatus"
          }
        },
        "type": "object",
        "required": [
          "status",
          "method",
          "final_status"
        ],
        "title": "VerificationResult"
      },
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "effective",
          "ineffective",
          "unknown"
        ],
        "title": "VerificationStatus"
      }
    }
  },
...
}